panchakarma_bot/requirements.txt
streamlit
sentence-transformers
faiss-cpu
google-generativeai
